dev-requirements.txt
beautifulsoup4
build
chartpress>=2.1
click
codecov
html5lib
jsonschema
jupyter-repo2docker>=2021.08.0
jupyter_packaging>=0.10.4,<2
jupyterhub
pytest
pytest-asyncio
pytest-cov
pytest-timeout
requests
ruamel.yaml>=0.17.30
